% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_venn_upset.R
\name{clean_venn_upset}
\alias{clean_venn_upset}
\title{Format gene lists for venn or upset plots}
\usage{
clean_venn_upset(model_result, models, variables, contrasts, intercept, random)
}
\value{
Data frame
}
\description{
Format gene lists for venn or upset plots
}
\keyword{internal}
