% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hm_clust.R
\name{get_hm_clust}
\alias{get_hm_clust}
\title{Get data frame of genes per cluster from ComplexHeatmap output}
\usage{
get_hm_clust(dat, hm, dimension)
}
\arguments{
\item{dat}{Ordered and scaled matrix of gene counts}

\item{hm}{ComplexHeatMap output specified with draw(h)}

\item{dimension}{Row or column clusters}
}
\value{
Data frame of genes per row/column, which row or column they owe to, and which cluster number
}
\description{
Get data frame of genes per cluster from ComplexHeatmap output
}
\examples{
library(ComplexHeatmap)
library(dplyr)
library(tidyr)
library(tibble)

hm.dat <- example.model$lme \%>\%
          select(gene, variable, pval) \%>\%
          pivot_wider(names_from = variable, values_from = pval) \%>\%
          column_to_rownames("gene") \%>\% as.matrix()
example.hm <- Heatmap(hm.dat, row_km=3, column_km=2)
example.hm <- draw(example.hm)

row_clust <- get_hm_clust(dat = hm.dat,
                        hm = example.hm, dimension = "row")
column_clust <- get_hm_clust(dat = hm.dat,
                        hm = example.hm, dimension = "col")
}
