% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corr_plot_matrix.R
\name{make_corr_plot_matrix}
\alias{make_corr_plot_matrix}
\title{Make Correlation Plot Matrix}
\usage{
make_corr_plot_matrix(
  dat = NULL,
  features_plot_square = NULL,
  features_plot_x = NULL,
  features_plot_y = NULL,
  lib = "libID",
  group_col = NULL,
  bp_groupname_rotate = FALSE,
  color_palette_custom = NULL,
  fontsize_labels = 25,
  labels_wrapwidth = 10,
  point_size = 1
)
}
\arguments{
\item{dat}{Dat object with sample-level metadata contained in dat$targets and expression data contained in dat$E}

\item{features_plot_square}{Vector of strings naming features to be plotted along both axes. Will create a square matrix of plots: points in the lower triangle, lines in the upper triangle. Boxplots of groupwise or overall median & IQR will be included along the vertical axis. Please provide only features_plot_square OR features_plot_x and features_plot_y.}

\item{features_plot_x}{Vector of strings naming features to be plotted along the horizontal axis. Will create two rectangular matrices of plots: points and lines. Boxplots of groupwise or overall median & IQR will be included along the horizontal axis. Please provide only features_plot_square OR features_plot_x and features_plot_y.}

\item{features_plot_y}{Vector of strings naming features to be plotted along the vertical axis. Will create two rectangular matrices of plots: points and lines. Boxplots of groupwise or overall median & IQR will be included along the vertical axis. Please provide only features_plot_square OR features_plot_x and features_plot_y.}

\item{lib}{String indicating name of column with unique library identifier in dat$targets (should match colnames in dat$E). Default is "libID"}

\item{group_col}{String indicating name of column with sample groups to be plotted from dat$targets. Can be left blank.}

\item{bp_groupname_rotate}{T/F: Do you want to rotate the group names on the x axis of the summary boxplots? Keeps them from running into each other if they are more than a few characters long.}

\item{color_palette_custom}{Custom color plot named vector, names are values in "group_col" column in dat$targets. otherwise ggplot default used.}

\item{fontsize_labels}{font size of feature labels along sides}

\item{labels_wrapwidth}{wrap width of feature labels along sides}

\item{point_size}{How big to make the points}
}
\value{
list of ggplot2 objects
}
\description{
Make Correlation Plot Matrix
}
\examples{
example.voom <- kimma::example.voom
make_corr_plot_matrix(dat = example.voom,
                      features_plot_square = rownames(example.voom$genes)[c(1:5)])

}
