% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_string.R
\name{map_string}
\alias{map_string}
\title{Map genes to string network}
\usage{
map_string(genes, version = 11.5, score_threshold = 400, species = "human")
}
\arguments{
\item{genes}{Character vector of HGNC symbols or ENSEMBL ID of genes}

\item{version}{Numeric version of STRING data base. Default 11.5 which is current as of 2022-03-30}

\item{score_threshold}{Numeric combined score threshold for edge inclusion in network. Default it 400 which is STRING defined "medium". 0 to 1000 allowed}

\item{species}{Character string of "human" or "mouse"}
}
\value{
List containing "subgraph" and "map"
}
\description{
To be used in conjunction with plot_string
}
\examples{
#See plot_string for full example
}
