% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrich.R
\name{plot_enrich}
\alias{plot_enrich}
\title{Plot k/K for hypergeo enrichment}
\usage{
plot_enrich(
  enrich,
  fdr_cutoff = 0.2,
  fdr_colors = c(0.01, 0.05, 0.1, 0.2),
  show_overlap = TRUE,
  fdr.cutoff = NULL,
  fdr.colors = NULL,
  show.overlap = NULL
)
}
\arguments{
\item{enrich}{Data frame output by SEARchways::BIGprofiler or SEARchways::BIGenrichr}

\item{fdr_cutoff}{Numeric. Maximum FDR to plot. Default is 0.2}

\item{fdr_colors}{Numeric vector. Cutoffs for color groups. Default is c(0.01, 0.05, 0.1, 0.2)}

\item{show_overlap}{Logical if should show overlap across all facets even if some missing (TRUE) or give each facet it's own axis labels (FALSE). Default is TRUE}

\item{fdr.cutoff}{Deprecated form of fdr_cutoff}

\item{fdr.colors}{Deprecated form of fdr_colors}

\item{show.overlap}{Deprecated form of show_overlap}
}
\value{
ggplot2 object
}
\description{
Plot k/K for hypergeo enrichment
}
\examples{
enrich <- data.frame("pathway" = c("pathway1", "pathway2", "pathway3"))
enrich$`k/K` = c(0.1, 0.1, 0.1)
enrich$genes = list(c("gene1", "gene2", "gene3"), c("gene3"), c("gene1", "gene2"))
enrich$FDR = c(0.01, 0.05, 0.2)
enrich$n_pathway_genes = c(10, 100, 1000)

#Plot
plot_enrich(enrich, fdr_cutoff = 0.5, fdr_colors = c(0.05, 0.5))
}
