% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrich2.R
\name{plot_enrich2}
\alias{plot_enrich2}
\title{Plotting function for hypergeometric enrichment}
\usage{
plot_enrich2(
  df = NULL,
  fdr_cutoff = 0.2,
  pathway_col = "pathway",
  gene_col = "genes",
  ratio_col = "k/K",
  fdr_col = "FDR",
  gssize_col = "n_pathway_genes",
  y_grouping_method = "fdr",
  include_gssize = FALSE,
  chart_style = "lollipop",
  fdr_binned_colors = c(0.001, 0.01, 0.05, 0.1, 0.2, 0.5, 0.99),
  fdr_continuous = FALSE,
  dot_sig_cutoff = NULL,
  dot_groupcol = "group"
)
}
\arguments{
\item{df}{Data frame output by SEARchways::BIGprofiler or SEARchways::flexEnrich}

\item{fdr_cutoff}{Numeric. Maximum FDR to plot. Default is 0.2.}

\item{pathway_col}{Character string Column name for df column containing gene set names.}

\item{gene_col}{Character string. Column name for df column containing gene lists names. Required if "y_grouping_method == 'hclust'".}

\item{ratio_col}{Character string Column name for df column containing k/K ratios.}

\item{fdr_col}{Character string. Column name for df column containing corrected p-values to plot.}

\item{gssize_col}{Character string. Column name for df column containing gene set sizes.}

\item{y_grouping_method}{Character string. Method for grouping gene sets along the y-axis. "hclust" (based on gene membership), "overlap_size", "gs_size", "ratio", "fdr", and "input". Default is "fdr".}

\item{include_gssize}{Boolean. Whether or not to include a column of gene set sizes to the left of the enrichment plot. Default is FALSE.}

\item{chart_style}{Character string. Options are "lollipop", "bar", and "dot". Default is "lollipop".}

\item{fdr_binned_colors}{Numeric vector. Cutoffs for binned FDR value color groups for lollipop plots. Default is c(0.001, 0.01, 0.05, 0.1, 0.2, 0.5, 0.99).}

\item{fdr_continuous}{Boolean. Whether or not to color points in a continuous fashions by a negative log transformed FDR value. Only applies to lollipop plots and dotplots.}

\item{dot_sig_cutoff}{Numeric. What cutoff to use for applying a black outline to dotplot dots}

\item{dot_groupcol}{Character string. Name of the column used to group enrichment along the x axis for dotplot.}
}
\value{
ggplot2 object
}
\description{
Plotting function for hypergeometric enrichment
}
\examples{
library(SEARchways)
library(dplyr)
#Run enrichment
gene_list <- list(HRV1 = names(example.gene.list[[1]]),
                  HRV2 = names(example.gene.list[[2]]))
enrich <- flexEnrich(gene_list, ID="ENSEMBL", collection="H")

#Plot
plot_enrich2(enrich, fdr_cutoff = 0.5)
}
