% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Model fit comparisons}
\usage{
plot_fit(model_result, model_result_y = NULL, x, y, metric = NULL)
}
\arguments{
\item{model_result}{List of data frames output by kimma::kmFit(). Must contain both x and y models if model_result_y not provided.}

\item{model_result_y}{List of data frame output by kimma::kmFit()}

\item{x}{Character string of model to plot on x-axis. Must match object names in model_result. For example, "lm", "lme", "lmekin"}

\item{y}{Character string of model to plot on y-axis. Must match object names in model_result. For example, "lm", "lme", "lmekin"}

\item{metric}{Character string of metric to plot. For example, "sigma", "AIC", "BIC", "Rsq", "}
}
\value{
ggplot object
}
\description{
Model fit comparisons
}
\examples{
plot_fit(model_result, x="lme", y="lmekin")
}
