% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA colored by variables of interest}
\usage{
plot_pca(dat, meta = NULL, vars, transform_logCPM = FALSE)
}
\arguments{
\item{dat}{Data frame, edgeR DGEList, or limma EList object containing gene counts in libraries}

\item{meta}{Data frame containing meta data with vars. Only needed if dat is a counts table and not an edgeR or limma object}

\item{vars}{Character vector of variables to color PCA by}

\item{transform_logCPM}{Logical if should convert counts to log counts per million}
}
\value{
List of ggplot objects
}
\description{
Plot PCA colored by variables of interest
}
