% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca2.R
\name{plot_pca2}
\alias{plot_pca2}
\title{Plot PCA colored by variables (does not calculate PCA)}
\usage{
plot_pca2(
  dat,
  vars,
  PCx = 1,
  PCy = 2,
  scale = FALSE,
  outlier_sd = 3,
  outlier_group = NULL,
  libraryID = "libID"
)
}
\arguments{
\item{dat}{edgeR DGEList, limma EList, or list object containing gene counts in libraries, metadata, and PCA calculations from calculate_pca()}

\item{vars}{Character vector of variables to color PCA by}

\item{PCx}{Numeric value for PC to plot on x-axis. Default is 1}

\item{PCy}{Numeric value for PC to plot on y-axis. Default is 2}

\item{scale}{Logical if you want to use scaled PCA values or not. see stats::prcomp for details. Default is FALSE}

\item{outlier_sd}{Numeric. If vars includes "outlier", statistical outliers are determined and colored based on this standard deviation along PC1 and PC2.}

\item{outlier_group}{Character string in which to group sd calculations}

\item{libraryID}{Character of variable name to match dat meta data frames}
}
\value{
List of ggplot objects
}
\description{
Plot PCA colored by variables (does not calculate PCA)
}
\examples{
plot_pca(calculate_pca(kimma::example.voom), var=c("virus","outlier"))
plot_pca(calculate_pca(kimma::example.voom), var=c("virus","outlier"), PCx=1, PCy=3)
}
