% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_hm.R
\name{plot_pca_hm}
\alias{plot_pca_hm}
\title{Plot PCA Correlation Heatmap (does not calculate PCA)}
\usage{
plot_pca_hm(
  dat,
  vars,
  scale = FALSE,
  corr_type = "pearson",
  pct_cutoff = 0.01,
  flip_axes = FALSE,
  show_signif_p = TRUE,
  signif_p_cutoff = 0.05,
  libraryID = "libID"
)
}
\arguments{
\item{dat}{edgeR DGEList, limma EList, or list object containing gene counts in libraries, metadata, and PCA calculations from calculate_pca()}

\item{vars}{Character vector of variables to color PCA by}

\item{scale}{Logical if should scale variance in PCA calculation see stats::prcomp for details. Default is FALSE}

\item{corr_type}{Character of correlation metric to use. Options: pearson,  spearman, or kendall. Default is pearson}

\item{pct_cutoff}{Numeric cutoff value for percent explained by PC (0 to 1). Default is 0.01 (1\%)}

\item{flip_axes}{Logical whether to swap x and y axes. Default is FALSE}

\item{show_signif_p}{Logical whether to show correlation significance. Default is TRUE}

\item{signif_p_cutoff}{Numeric correlation p-value cutoff (0 to 1). Default is 0.05}

\item{libraryID}{Character of variable name to match dat meta data frames}
}
\value{
ggplot2 object
}
\description{
Plot PCA Correlation Heatmap (does not calculate PCA)
}
\examples{
dat_PCA <- calculate_pca(kimma::example.voom,scale=TRUE)
plot_pca_hm(dat_PCA,vars = c("median_cv_coverage","lib.size"),scale=TRUE,flip_axes = TRUE)
}
