% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_string.R
\name{plot_string}
\alias{plot_string}
\title{Plot STRING network of genes colored by enriched pathways}
\usage{
plot_string(
  map,
  discard = "none",
  layout = "fr",
  enrichment = NULL,
  overlap = 2,
  FDR = 0.2,
  ID = c("SYMBOLs", "ENTREZIDs", "ENSEMBLIDs"),
  colors = NULL,
  text_size = 2,
  node_size = 1
)
}
\arguments{
\item{map}{List output by map_string}

\item{discard}{Character string identifying genes to remove from the network. Can be "none" (Default), "orphan" (genes with 0 connections in the network), or "orphan.unenrich" (genes with 0 connections and no significant enrichment colors)}

\item{layout}{Character string for network layout algorithm. See options in igraph::layout_with_}

\item{enrichment}{Data frame output by hypergeometric enrichment}

\item{overlap}{Numeric minimum of significant genes in terms to be used as colors}

\item{FDR}{Numeric maximum FDR of terms to be used as colors}

\item{ID}{Character string defining gene ID column in enrichment that matches genes used in map_string. Can be "SYMBOLs" (Default), "ENTREZIDs", "ENSEMBLIDs"}

\item{colors}{Character vector of custom colors to use. Must be at least a long as total significant terms plus 1 for the "none" group}

\item{text_size}{Numeric size of gene labels on network nodes. Default of 2}

\item{node_size}{Numeric size of network nodes. Default of 1}
}
\value{
ggplot STRING network object
}
\description{
To be used in conjunction with map_string
}
\examples{
map <- map_string(genes = c("WNT5B","CCND2","FZD1","CTNNB1","DKK1","TCF7"),
                       version = 11.5, score_threshold = 400)
plot_string(map)
plot_string(map, enrichment = BIGpicture::enrichment, ID ="SYMBOLs")
}
