% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset_genes.R
\name{plot_upset_genes}
\alias{plot_upset_genes}
\title{Upset plots of significant genes}
\usage{
plot_upset_genes(
  model_result,
  models = NULL,
  variables = NULL,
  contrasts = NULL,
  intercept = FALSE,
  random = FALSE,
  return_genes = FALSE,
  fdr_cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  return.genes = NULL,
  fdr.cutoff = NULL
)
}
\arguments{
\item{model_result}{NAMED list of lists output by kimma::kmFit() such as list("model1"=model_result1, "model2"=model_result2)}

\item{models}{Character vector of model(s) to plot. Must match object names in model_result. For example, "lm", "lme", "lmerel", "limma" or similar with ".contrast"}

\item{variables}{Character vector of variables in model_result to include in plots}

\item{contrasts}{Character vector of contrasts in model_result to include in plots. Format is c("contrast_lvl - contrast_ref", "..."). Only applicable if model name includes 'contrast'}

\item{intercept}{Logical if should include the intercept variable. Default is FALSE}

\item{random}{Logical if should include random effect variable(s). Default is FALSE}

\item{return_genes}{Logical if should return data frame of genes in upsets. Default is FALSE}

\item{fdr_cutoff}{Numeric vector of FDR cutoffs to assess. One upset per FDR value}

\item{return.genes}{Deprecated form of return_genes}

\item{fdr.cutoff}{Deprecated form of fdr_cutoff}
}
\value{
List with 1 each for each FDR cutoff of (1) upset diagram ggplot object and (2) data frame of genes in upset
}
\description{
Upset plots of significant genes
}
\examples{
# A single model, multiple variables
upset.result <- plot_upset_genes(model_result = list("example.model" = example.model),
    models = "lme", return_genes = TRUE,
    fdr_cutoff = c(0.05,0.5))
#plot all upset
patchwork::wrap_plots(upset.result[["upset"]])
#Plot 1 upset
upset.result[["upset"]][["0.05"]]
#see genes in intersections because return_genes = TRUE
upset.result[["gene"]]

# Multiple models, subset of variables
plot_upset_genes(model_result = list("example.model" = example.model),
    models = c("lme","lmerel"),
    variables = c("virus","virus:asthma"),
    fdr_cutoff = c(0.05))
# Multiple models in two different objects, subset of variables
plot_upset_genes(model_result = list("model1" = example.model[c(1,3,5)],
                                     "model2" = example.model[c(2,4,6)]),
    models = c("lme","lmerel"),
    variables = c("virus","virus:asthma"),
    fdr_cutoff = c(0.05))

# Contrasts
plot_upset_genes(model_result = list("example.model" = example.model),
    models = "lme.contrast",
    contrasts = c("HRV asthma - none asthma",
                  "HRV healthy - none healthy"),
    fdr_cutoff = c(0.5))
plot_upset_genes(model_result = list("model1" = example.model[c(1,3,5)],
                                     "model2" = example.model[c(2,4,6)]),
    models = c("lme.contrast","lmerel.contrast"),
    contrasts = c("HRV asthma - none asthma",
                  "HRV healthy - none healthy"),
    fdr_cutoff = c(0.5))
}
