% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of differentially expressed genes}
\usage{
plot_volcano(
  model_result,
  model,
  variables = NULL,
  x = "estimate",
  y = "FDR",
  estimate_cutoff = NULL,
  fdr_cutoff = NULL,
  label = NULL,
  genes = NULL,
  genes_label = NULL,
  x.cutoff = NULL,
  y.cutoff = NULL
)
}
\arguments{
\item{model_result}{List of data frames output by kimma::kmFit()}

\item{model}{Character string of model to plot. Must match object names in model_result. For example, "lm", "lme", "lmerel"}

\item{variables}{Character vector of variables in model_result to include. Default is all variables in model}

\item{x}{Character string of variable to plot on x-axis. Default is "estimate"}

\item{y}{Character string of variable to plot on y-axis. Default is "FDR"}

\item{estimate_cutoff}{Numeric. Optional estimate (fold change or slope) cutoff for color and/or labeling}

\item{fdr_cutoff}{Numeric. Optional FDR cutoff for color and/or labeling}

\item{label}{Character or numeric. If "all", all significant genes as defined
by estimate_cutoff and fdr_cutoff are labels with their HGNC symbol. If numeric, that number of most significant genes are labeled.}

\item{genes}{Data frame with gene metadata for labeling points (optional). If not provided, the gene column in the model_result is used}

\item{genes_label}{Character string of variable in genes to label with. Required if provide genes parameter}

\item{x.cutoff}{Superseded by estimate_cutoff}

\item{y.cutoff}{Superseded by fdr_cutoff}
}
\value{
ggplot object
}
\description{
Volcano plot of differentially expressed genes
}
\examples{
plot_volcano(example_model, model = "lme")
plot_volcano(example_model, model = "lme", variables = "virus", y = "pval")

#Select specific variables and cutoffs
plot_volcano(example_model, model = "lme", variables = c("virus","asthma"),
             estimate_cutoff = 0.5, fdr_cutoff = 0.05, label = 2)

#Label top genes
plot_volcano(example_model, model = "lme", variables = "virus",
             fdr_cutoff = 1E-20, label = "all")
plot_volcano(example_model, model = "lme", variables = "virus",
             fdr_cutoff = 1E-20, label = "all",
             genes = kimma::example.voom$genes, genes_label = "hgnc_symbol")

#Plot contrasts
plot_volcano(example_model, model = "lme.contrast",
             variables = "virus*asthma",
             fdr_cutoff = 0.05, label="all")
}
