% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_plot.R
\name{sigma_plot}
\alias{sigma_plot}
\title{Model fit comparisons using sigma estimate}
\usage{
sigma_plot(model_result, x, y)
}
\arguments{
\item{model_result}{List of data frames output kimma::kmFit()}

\item{x}{Character string of model to plot on x-axis. Must match object names in model_result. For example, "lm", "lme", "lmekin"}

\item{y}{Character string of model to plot on y-axis. Must match object names in model_result. For example, "lm", "lme", "lmekin"}
}
\value{
ggplot object
}
\description{
Model fit comparisons using sigma estimate
}
\examples{
sigma_plot(model_results, x="lme", y="lmekin")
}
