% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genes.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot boxplots and dotplots for genes and variables of interest}
\usage{
plot_genes(
  dat = NULL,
  counts = NULL,
  meta = NULL,
  genes = NULL,
  fdr = NULL,
  libraryID = "libID",
  geneID = "geneName",
  subset_genes = NULL,
  variables,
  colorID = NULL,
  processors = NULL,
  subset.genes = NULL
)
}
\arguments{
\item{dat}{DGEList or EList object with expression data (counts, E), sample metadata (samples, targets), and gene annotation (genes)}

\item{counts}{If dat not provided. Data frame of gene expression. Genes are rows, samples are columns. geneID must be rownames or first column}

\item{meta}{If dat not provided. Data frame of sample meta data with samples as rows.}

\item{genes}{Optional if dat not provided. Data frame of gene annotation with genes as rows.}

\item{fdr}{Optional. Model results output by kmFit( ). Used to label significant differences}

\item{libraryID}{Character string of variable name to use when combining expression and sample data}

\item{geneID}{Character string of variable name to use when combining expression and gene data}

\item{subset_genes}{Optional. Character vector of genes to plot. Must match names in geneID column. If not provided, all genes are plotted.}

\item{variables}{Character vector of variable names to include in plot. Variables can be character, factor, or numeric. One two-variable interaction term allowed}

\item{colorID}{Optional. Character string for variable to color point by}

\item{processors}{Numeric processors to run in parallel. Default is 2 less than the total available}

\item{subset.genes}{Deprecated form of subset_genes}
}
\value{
List of ggplot objects
}
\description{
Plot boxplots and dotplots for genes and variables of interest
}
\examples{
#Data from kimma
example.voom <- kimma::example.voom

subset_genes <- c("ENSG00000250479","ENSG00000250510","ENSG00000255823")

plot_genes(dat = example.voom, fdr = example.model$lmerel,
     subset_genes = subset_genes, geneID="geneName",
     variables = c("virus*asthma", "lib.size"), colorID = "virus")
}
