% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot GSEA normalized enrichment}
\usage{
plot_gsea(
  gsea,
  fdr.cutoff = 0.2,
  fdr.colors = c(0.01, 0.05, 0.1, 0.2),
  show.overlap = TRUE
)
}
\arguments{
\item{gsea}{Data frame output by SEARchways::BIGsea including pathway, FDR, and NES}

\item{fdr.cutoff}{Numeric. Maximum FDR to plot. Default is 0.2}

\item{fdr.colors}{Numeric vector. Cutoffs for color groups. Default is c(0.01, 0.05, 0.1, 0.2)}

\item{show.overlap}{Logical if should show overlap across all facets even if some missing (TRUE) or give each facet it's own axis labels (FALSE). Default is TRUE}
}
\value{
ggplot2 object
}
\description{
Plot GSEA normalized enrichment
}
\examples{
library(SEARchways)
library(dplyr)
#Get fold change information from example model
genes.FC <- example.model$lmerel \%>\%
            filter(variable == "virus") \%>\%
            select(variable, gene, estimate)
#Run GSEA
example_gsea <- SEARchways::BIGsea(gene_df = genes.FC, category = "H", ID = "ENSEMBL")

#Plot
plot_gsea(example_gsea, fdr.cutoff = 0.5)
}
