% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot GSEA normalized enrichment}
\usage{
plot_gsea(
  gsea,
  fdr_cutoff = 0.2,
  fdr_colors = c(0.01, 0.05, 0.1, 0.2),
  show_overlap = TRUE,
  fdr.cutoff = NULL,
  fdr.colors = NULL,
  show.overlap = NULL
)
}
\arguments{
\item{gsea}{Data frame output by SEARchways::BIGsea including pathway, FDR, and NES}

\item{fdr_cutoff}{Numeric. Maximum FDR to plot. Default is 0.2}

\item{fdr_colors}{Numeric vector. Cutoffs for color groups. Default is c(0.01, 0.05, 0.1, 0.2)}

\item{show_overlap}{Logical if should show overlap across all facets even if some missing (TRUE) or give each facet it's own axis labels (FALSE). Default is TRUE}

\item{fdr.cutoff}{Deprecated form of fdr_cutoff}

\item{fdr.colors}{NDeprecated form of fdr_colors}

\item{show.overlap}{Deprecated form of show_overlap}
}
\value{
ggplot2 object
}
\description{
Plot GSEA normalized enrichment
}
\examples{
library(SEARchways)
library(dplyr)
#Get fold change information from example model
genes.FC <- example.model$lmerel \%>\%
            filter(variable \%in\% c("virus", "asthma")) \%>\%
            select(variable, gene, estimate)
#Run GSEA
example_gsea <- SEARchways::BIGsea(gene_df = genes.FC, category = "H", ID = "ENSEMBL")

#Plot
plot_gsea(example_gsea, fdr_cutoff = 0.5,
         fdr_colors = c(0.1, 0.5))
}
