% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA colored by variables of interest}
\usage{
plot_pca(
  dat,
  meta = NULL,
  vars,
  PCx = 1,
  PCy = 2,
  scale = FALSE,
  outlier_sd = 3,
  outlier_group = NULL,
  transform_logCPM = FALSE,
  libraryID = "libID"
)
}
\arguments{
\item{dat}{Data frame, edgeR DGEList, or limma EList object containing gene counts in libraries}

\item{meta}{Data frame containing meta data with vars. Only needed if dat is a counts table and not an edgeR or limma object}

\item{vars}{Character vector of variables to color PCA by}

\item{PCx}{Numeric value for PC to plot on x-axis. Default it 1}

\item{PCy}{Numeric value for PC to plot on y-axis. Default it 2}

\item{scale}{Logical if should scale variance in PCA calculation see stats::prcomp for details. Default is FALSE}

\item{outlier_sd}{Numeric. If vars includes "outlier", statistical outliers are determined and colored based on this standard deviation along PC1 and PC2.}

\item{outlier_group}{Character string in which to group sd calculations}

\item{transform_logCPM}{Logical if should convert counts to log counts per million}

\item{libraryID}{Character of variable name to match dat meta data frames}
}
\value{
List of ggplot objects
}
\description{
Plot PCA colored by variables of interest
}
\examples{
plot_pca(kimma::example.voom, var=c("virus","outlier"))
plot_pca(kimma::example.voom, var=c("virus","outlier"), PCx=1, PCy=3)
}
