% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_string.R
\name{plot_string}
\alias{plot_string}
\title{Plot STRING network of genes colored by enriched pathways}
\usage{
plot_string(
  map,
  discard = "none",
  layout = "fr",
  enrichment = NULL,
  overlap = 2,
  fdr.cutoff = 0.2,
  colors = NULL,
  text_size = 2,
  node_size = 1
)
}
\arguments{
\item{map}{List output by map_string}

\item{discard}{Character string identifying genes to remove from the network. Can be "none" (Default), "orphan" (genes with 0 connections in the network), or "orphan.unenrich" (genes with 0 connections and no significant enrichment colors)}

\item{layout}{Character string for network layout algorithm. See options in igraph::layout_with_}

\item{enrichment}{Data frame output by hypergeometric enrichment}

\item{overlap}{Numeric minimum of significant genes in terms to be used as colors}

\item{fdr.cutoff}{Numeric maximum FDR of terms to be used as colors}

\item{colors}{Character vector of custom colors to use. Must be at least a long as total significant terms plus 1 for the "none" group}

\item{text_size}{Numeric size of gene labels on network nodes. Default of 2}

\item{node_size}{Numeric size of network nodes. Default of 1}
}
\value{
ggplot STRING network object
}
\description{
To be used in conjunction with map_string
}
\examples{
map <- map_string(genes = c("MTHFD2","ISOC1","IL2RB","SAMHD1","NAMPT","NOD1",
                            "NFKB1","IFIT3","ZBP1","IL15RA","SP110","ITGB7",
                            "SERPING1","B2M","CXCL11","USP18","MAPKAPK2","DKK1"),
                 version = 11.5, score_threshold = 400)
plot_string(map)

# Add enrichment colors
library(dplyr)
library(SEARchways)
genes.OI <- example_model$lmerel \%>\%
            filter(variable == "virus" & FDR < 0.2) \%>\%
            distinct(variable, hgnc_symbol)
example_enrich <- BIGprofiler(gene_df = genes.OI, category = "H")

plot_string(map, enrichment = example_enrich, fdr.cutoff=0.2)

# Add GSEA colors
genes.FC <- example_model$lmerel \%>\%
            filter(variable == "virus") \%>\%
            select(variable, hgnc_symbol, estimate)
example_gsea <- BIGsea(gene_df = genes.FC, category = "H")

plot_string(map, enrichment = example_gsea, fdr.cutoff=0.2)
}
