% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset_genes.R
\name{plot_upset_genes}
\alias{plot_upset_genes}
\title{Upset plots of significant genes}
\usage{
plot_upset_genes(
  model_result,
  models = NULL,
  variables = NULL,
  contrasts = NULL,
  intercept = FALSE,
  random = FALSE,
  return.genes = FALSE,
  fdr.cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5)
)
}
\arguments{
\item{model_result}{NAMED list of lists output by kimma::kmFit() such as list("model1"=model_result1, "model2"=model_result2)}

\item{models}{Character vector of model(s) to plot. Must match object names in model_result. For example, "lm", "lme", "lmerel", "limma"}

\item{variables}{Character vector of variables in model_result to include in plots}

\item{contrasts}{Character vector of contrasts in model_result to include in plots. Format is c("contrast_lvl - contrast_ref", "..."). Only applicable if model name includes 'contrast'}

\item{intercept}{Logical if should include the intercept variable. Default is FALSE}

\item{random}{Logical if should include random effect variable(s). Default is FALSE}

\item{return.genes}{Logical if should return data frame of genes in upsets. Default is FALSE}

\item{fdr.cutoff}{Numeric vector of FDR cutoffs to assess. One upset per FDR value}
}
\value{
List with 1 each for each FDR cutoff of (1) upset diagram ggplot object and (2) data frame of genes in upset
}
\description{
Upset plots of significant genes
}
\examples{
# A single model, multiple variables
upset.result <- plot_upset_genes(model_result = list("example_model" = example_model),
    models = "lme", return.genes = TRUE,
    fdr.cutoff = c(0.05,0.5))
#plot all upset
patchwork::wrap_plots(upset.result[["upset"]])
#Plot 1 upset
upset.result[["upset"]][["0.05"]]
#see genes in intersections
upset.result[["gene"]]

# Multiple models, subset of variables
model1 <- list("lme" = example_model$lme)
model2 <- list("lmerel" = example_model$lmerel)
plot_upset_genes(list("lme"=model1, "lmerel"=model2),
    variables = c("virus","virus:asthma"),
    fdr.cutoff = c(0.05))

# Contrasts
model1 <- list("lme" = example_model$lme.contrast)
model2 <- list("lmerel" = example_model$lmerel.contrast)
plot_upset_genes(model_result = list("lme"=model1, "lmerel"=model2),
    contrasts = c("HRV asthma - none asthma"),
    fdr.cutoff = c(0.4))
}
