% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn_genes.R
\name{plot_venn_genes}
\alias{plot_venn_genes}
\title{Venn diagrams of significant genes}
\usage{
plot_venn_genes(
  model_result,
  model,
  variables = NULL,
  contrasts = NULL,
  intercept = FALSE,
  random = FALSE,
  fdr.cutoff = c(0.05, 0.1, 0.2, 0.3, 0.4, 0.5)
)
}
\arguments{
\item{model_result}{List of data frames output by kimma::kmFit()}

\item{model}{Character string of model to plot. Must match object names in model_result. For example, "lm", "lme", "lmekin"}

\item{variables}{Character vector of variables in model_result to include in plots}

\item{contrasts}{Character vector of contrasts in model_result to include in plots. Format is c("contrast_lvl - contrast_ref", "..."). Only applicable if model name includes 'contrast'}

\item{intercept}{Logical if should include the intercept variable. Default is FALSE}

\item{random}{Logical if should include random effect variable(s). Default is FALSE}

\item{fdr.cutoff}{Numeric vector of FDR cutoffs to assess. One venn per FDR value}
}
\value{
List of ggplot objects
}
\description{
Venn diagrams of significant genes
}
\examples{
plot_venn_genes(example_model, model = "lme", fdr.cutoff = c(0.05,0.5))
}
